Renamed `solar_setbacks.py to the more appropriate `parcel_setbacks.py`
