+++ reVX/setbacks/__init__.py
@@ -5,4 +5,4 @@ Compute setbacks
 """
 from .wind_setbacks import (StructureWindSetbacks, RoadWindSetbacks,
                             RailWindSetbacks, TransmissionWindSetbacks)
-from .solar_setbacks import ParcelSetbacks
+from .parcel_setbacks import ParcelSetbacks
